+++ src/test/kotlin/com/atlassian/performance/tools/virtualusers/api/EntryPointIT.kt
@@ -4,16 +4,22 @@ import com.atlassian.performance.tools.infrastructure.api.database.MySqlDatabase
 import com.atlassian.performance.tools.infrastructure.api.dataset.Dataset
 import com.atlassian.performance.tools.infrastructure.api.dataset.HttpDatasetPackage
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraHomePackage
+import com.atlassian.performance.tools.jiraactions.api.ActionMetric
 import com.atlassian.performance.tools.virtualusers.ChromeContainer
 import com.atlassian.performance.tools.virtualusers.DockerJiraFormula
 import com.atlassian.performance.tools.virtualusers.SimpleScenario
 import com.atlassian.performance.tools.virtualusers.TestVuNode
+import com.atlassian.performance.tools.virtualusers.api.ActivityType.ACTING
+import com.atlassian.performance.tools.virtualusers.api.ActivityType.DIAGNOSING
+import com.atlassian.performance.tools.virtualusers.api.ActivityType.MYSTERY
+import com.atlassian.performance.tools.virtualusers.api.ActivityType.THROTTLING
 import com.atlassian.performance.tools.virtualusers.lib.infrastructure.Jperf424WorkaroundJswDistro
 import com.atlassian.performance.tools.virtualusers.lib.infrastructure.Jperf425WorkaroundMysqlDatabase
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.net.URI
 import java.time.Duration
+import java.util.stream.Stream
 import kotlin.streams.asSequence
 
 class EntryPointIT {
@@ -63,15 +69,35 @@ class EntryPointIT {
             .listResults()
             .last()
         val scenarioLabels = vuResult
-            .streamMetrics()
+            .streamScenarioMetrics()
             .map { it.label }
         assertThat(scenarioLabels).containsOnly("Log In", "View Issue")
+        val activityTypes = vuResult
+            .streamActivityMetrics()
+            .map { it.type }
+        assertThat(activityTypes).contains(ACTING, THROTTLING, DIAGNOSING).doesNotContain(MYSTERY)
+        val totalActivityTime = vuResult
+            .streamActivityMetrics()
+            .map { it.metric }
+            .sumDurations()
+        val unaccountedTime = desiredTotalTime - totalActivityTime
+        assertThat(unaccountedTime).isLessThan(Duration.ofSeconds(5))
         val totalMetricsTime = vuResult
-            .streamMetrics()
-            .map { it.duration }
+
+            .streamScenarioMetrics()
+            .sumDurations()
+        val totalActingTime = vuResult
+            .streamActivityMetrics()
+            .filter { it.type == ACTING }
+            .map { it.metric }
+            .sumDurations()
+        val unmeasuredActingTime = totalActingTime - totalMetricsTime
+        assertThat(unmeasuredActingTime).isLessThan(Duration.ofSeconds(1))
+    }
+
+    private fun Stream<ActionMetric>.sumDurations(): Duration {
+        return map { it.duration }
             .asSequence()
             .fold(Duration.ZERO) { a, b -> a + b }
-        val unaccountedTime = desiredTotalTime - totalMetricsTime
-        assertThat(unaccountedTime).isLessThan(Duration.ofSeconds(5))
     }
 }